app/requirements.txt
gradio>=5.0
pillow>=10.0
numpy>=1.26
scikit-learn>=1.4
